Normalized temperature (#55)

+++ src/llm_models/google.rs
@@ -82,7 +82,7 @@ impl LLMModel for GoogleModels {
         json_schema: &Value,
         function_call: bool,
         _max_tokens: &usize,
-        temperature: &u32,
+        temperature: &f32,
     ) -> serde_json::Value {
         //Prepare the 'messages' part of the body
         let base_instructions_json = json!({
